+++ src/fractal_uzh_converters/common/__init__.py
@@ -3,9 +3,14 @@
 from fractal_uzh_converters.common.image_in_plate_compute_task import (
     image_in_plate_compute_task,
 )
+from fractal_uzh_converters.common.single_image_compute_task import (
+    single_image_compute_task,
+)
 from fractal_uzh_converters.common.utils import (
     STANDARD_ROWS_NAMES,
     BaseAcquisitionModel,
+    HCSBaseAcquisitionModel,
+    SingleBaseAcquisitionModel,
     get_attributes_from_condition_table,
     parse_acquisitions,
 )
@@ -13,7 +18,10 @@ from fractal_uzh_converters.common.utils import (
 __all__ = [
     "STANDARD_ROWS_NAMES",
     "BaseAcquisitionModel",
+    "HCSBaseAcquisitionModel",
+    "SingleBaseAcquisitionModel",
     "get_attributes_from_condition_table",
     "image_in_plate_compute_task",
     "parse_acquisitions",
+    "single_image_compute_task",
 ]